Tx.Copy avoids cache pollution on Linux
Avoid overwhelming page cache with all of the database pages, by
immediately discarding the pages that were not already in RAM before
the Copy.

Keep track of what pages were already in RAM, so we don't lose the
cached that are part of the working set.

This allows us to benefit from any cached pages, and avoid O_DIRECT,
while minimizing performance impact of a database copy on a live
service.

Note that MADV_DONTNEED is merely advice to the kernel. On a host with
plenty of available RAM, on an mmapped source file with random
accesses, some or all of the pages may still remain in RAM. Only when
there is a shortage of available RAM for page cache, does madvise seem
to take full effect.